+++ R/QIM_report_UI.R
@@ -264,9 +264,44 @@ qim_reportCreator <- function(input, output, session, dMQIM) {
           value = 0
         )
 
-        if (measure_names[[1]] %in% input$report_qim_chosen) {
-          progress$inc(amount = 1, detail = "QIM 01 - Diabetes HbA1C")
-          qim01 <- dMQIM$report_qim_diabetes(
+        qim <- data.frame(
+          QIM = character(),
+          Age10 = numeric(),
+          Sex = character(),
+          Indigenous = character(),
+          DiabetesType = character(),
+          Measure = character(),
+          State = character(),
+          n = numeric(),
+          ProportionDemographic = numeric(),
+          DateFrom = character(),
+          DateTo = character()
+        )
+
+        # gets report from 'report_function'
+        # then adds extra 'DiabetesType' column if necessary
+        # adds more columns describing the measurement
+        # renames the 'result' column to 'State'
+        #
+        # called to create reports in a 'standard' format
+        # from the 10 QIM measures
+        getReport <- function(
+          report_function, # the dMQIM method to call
+          progress_detail, # the message to report in the 'progress' dialog
+          measure = NA,
+          # some (but not all) 'report_function' can choose the 'measure'
+          # if not required, then use 'NA'
+          require_type_diabetes,
+          # diabetes methods may give option to specify diabetes types
+          qim_name, measure_name,
+          # the overall name of the QIM, and the name of the measure
+          state_variable_name
+          # the name of the column in the return from 'report_function'
+          # which becomes the 'State' column
+        ) {
+          progress$inc(amount = 1, detail = progress_detail)
+
+          function_args <- list(
             contact = TRUE, date_from = date_from, date_to = date_to,
             min_contact = min_contact,
             # 'default' is clinician list chosen in right panel
@@ -274,345 +309,196 @@ qim_reportCreator <- function(input, output, session, dMQIM) {
             # no limitations on min/max date essentially
             contact_type = contact_type,
             demographic = c("Age10", "Sex", "Indigenous"),
-            measure = "HbA1C", type_diabetes = TRUE,
             ignoreOld = TRUE, lazy = FALSE, store = FALSE
-          ) %>>%
-            dplyr::mutate(QIM = "QIM 01",
-                          Measure = "HbA1C",
-                          DateFrom = as.character(date_from),
-                          DateTo = as.character(date_to)) %>>%
-            dplyr::rename(State = HbA1CDone,
-                          ProportionDemographic = Proportion_Demographic) %>>%
-            dplyr::select(QIM, Age10, Sex, Indigenous, DiabetesType,
-                          Measure, State, n, ProportionDemographic,
-                          DateFrom, DateTo)
+          )
+          if (!is.na(measure)) {
+            function_args <- append(
+              function_args,
+              list(measure = measure)
+            )
+          }
+          if (require_type_diabetes) {
+            function_args <- append(
+              function_args,
+              list(type_diabetes = TRUE)
+            )
+          }
+          qim_report <- do.call(report_function, function_args) %>>%
+            dplyr::mutate(
+              QIM = qim_name, Measure = measure_name
+            ) %>>% {
+              if (require_type_diabetes) {
+                . # 'DiabetesType' should already exit
+              } else {
+                dplyr::mutate(.,
+                  DiabetesType = "" # create a 'blank'
+                )
+              }
+            } %>>%
+            dplyr::rename(
+              State = !!state_variable_name,
+              ProportionDemographic = Proportion_Demographic
+            ) %>>%
+            dplyr::select(
+              QIM, Age10, Sex, Indigenous, DiabetesType,
+              Measure, State, n, ProportionDemographic
+            )
           # keep Age10, Sex, Indigenous, DiabetesType, HbA1CDone,
           # n, Proportion_Demographic
-          report_values(
-            rbind(
-              report_values(),
-              qim01
-            )
+
+          return(qim_report)
+        }
+
+        if (measure_names[[1]] %in% input$report_qim_chosen) {
+          qim01 <- getReport(
+            dMQIM$report_qim_diabetes,
+            progress_detail = "QIM 01 - Diabetes HbA1C",
+            measure = "HbA1C", require_type_diabetes = TRUE,
+            qim_name = "QIM 01", measure_name = "HbA1C",
+            state_variable_name = dplyr::quo(HbA1CDone)
           )
+          qim <- rbind(qim, qim01)
         }
 
         if (measure_names[[2]] %in% input$report_qim_chosen) {
-          progress$inc(amount = 1, detail = "QIM 02 - 15+ smoking")
-          qim02 <- dMQIM$report_qim_15plus(
-            contact = TRUE, date_from = date_from, date_to = date_to,
-            min_contact = min_contact,
-            # 'default' is clinician list chosen in right panel
-            min_date = as.Date("2000-01-01"), max_date = Sys.Date(),
-            # no limitations on min/max date essentially
-            contact_type = contact_type,
-            demographic = c("Age10", "Sex", "Indigenous"),
-            measure = "Smoking",
-            ignoreOld = TRUE, lazy = FALSE, store = FALSE
-          ) %>>%
-            dplyr::mutate(QIM = "QIM 02",
-                          Measure = "Smoking",
-                          DiabetesType = "",
-                          DateFrom = as.character(date_from),
-                          DateTo = as.character(date_to)) %>>%
-            dplyr::rename(State = SmokingStatus,
-                          ProportionDemographic = Proportion_Demographic) %>>%
-            dplyr::select(QIM, Age10, Sex, Indigenous, DiabetesType,
-                          Measure, State, n, ProportionDemographic,
-                          DateFrom, DateTo)
-          # keep Age10, Sex, Indigenous, DiabetesType, HbA1CDone,
-          # n, Proportion_Demographic
-          report_values(
-            rbind(
-              report_values(),
-              qim02
-            )
+          qim02 <- getReport(
+            dMQIM$report_qim_15plus,
+            progress_detail = "QIM 02 - 15+ smoking",
+            measure = "Smoking", require_type_diabetes = FALSE,
+            qim_name = "QIM 02", measure_name = "Smoking",
+            state_variable_name = dplyr::quo(SmokingStatus)
           )
+          qim <- rbind(qim, qim02)
         }
 
         if (measure_names[[3]] %in% input$report_qim_chosen) {
-          progress$inc(amount = 1, detail = "QIM 03 - 15+ BMI Class")
-          qim03 <- dMQIM$report_qim_15plus(
-            contact = TRUE, date_from = date_from, date_to = date_to,
-            min_contact = min_contact,
-            # 'default' is clinician list chosen in right panel
-            min_date = as.Date("2000-01-01"), max_date = Sys.Date(),
-            # no limitations on min/max date essentially
-            contact_type = contact_type,
-            demographic = c("Age10", "Sex", "Indigenous"),
-            measure = "Weight",
-            ignoreOld = TRUE, lazy = FALSE, store = FALSE
-          ) %>>%
-            dplyr::mutate(QIM = "QIM 03",
-                          Measure = "BMIClass",
-                          DiabetesType = "",
-                          DateFrom = as.character(date_from),
-                          DateTo = as.character(date_to)) %>>%
-            dplyr::rename(State = BMIClass,
-                          ProportionDemographic = Proportion_Demographic) %>>%
-            dplyr::select(QIM, Age10, Sex, Indigenous, DiabetesType,
-                          Measure, State, n, ProportionDemographic,
-                          DateFrom, DateTo)
-          # keep Age10, Sex, Indigenous, DiabetesType, HbA1CDone,
-          # n, Proportion_Demographic
-          report_values(
-            rbind(
-              report_values(),
-              qim03
-            )
+          qim03 <- getReport(
+            dMQIM$report_qim_15plus,
+            progress_detail = "QIM 03 - 15+ BMI Class",
+            measure = "Weight", require_type_diabetes = FALSE,
+            qim_name = "QIM 03", measure_name = "BMIClass",
+            state_variable_name = dplyr::quo(BMIClass)
           )
+          qim <- rbind(qim, qim03)
         }
 
         if (measure_names[[4]] %in% input$report_qim_chosen) {
-          progress$inc(amount = 1, detail = "QIM 04 - 65+ Influenza")
-          qim04 <- dMQIM$report_qim_65plus(
-            contact = TRUE, date_from = date_from, date_to = date_to,
-            min_contact = min_contact,
-            # 'default' is clinician list chosen in right panel
-            min_date = as.Date("2000-01-01"), max_date = Sys.Date(),
-            # no limitations on min/max date essentially
-            contact_type = contact_type,
-            demographic = c("Age10", "Sex", "Indigenous"),
-            ignoreOld = TRUE, lazy = FALSE, store = FALSE
-          ) %>>%
-            dplyr::mutate(QIM = "QIM 04",
-                          Measure = "InfluenzaDone",
-                          DiabetesType = "",
-                          DateFrom = as.character(date_from),
-                          DateTo = as.character(date_to)) %>>%
-            dplyr::rename(State = InfluenzaDone,
-                          ProportionDemographic = Proportion_Demographic) %>>%
-            dplyr::select(QIM, Age10, Sex, Indigenous, DiabetesType,
-                          Measure, State, n, ProportionDemographic,
-                          DateFrom, DateTo)
-          # keep Age10, Sex, Indigenous, DiabetesType, HbA1CDone,
-          # n, Proportion_Demographic
-          report_values(
-            rbind(
-              report_values(),
-              qim04
-            )
+          qim04 <- getReport(
+            dMQIM$report_qim_65plus,
+            progress_detail = "QIM 04 - 64+ Influenza",
+            measure = NA, require_type_diabetes = FALSE,
+            qim_name = "QIM 04", measure_name = "InfluenzaDone",
+            state_variable_name = dplyr::quo(InfluenzaDone)
           )
+          qim <- rbind(qim, qim04)
         }
+
         if (measure_names[[5]] %in% input$report_qim_chosen) {
-          progress$inc(amount = 1, detail = "QIM 05 - Diabetes Influenza")
-          qim05 <- dMQIM$report_qim_diabetes(
-            contact = TRUE, date_from = date_from, date_to = date_to,
-            min_contact = min_contact,
-            # 'default' is clinician list chosen in right panel
-            min_date = as.Date("2000-01-01"), max_date = Sys.Date(),
-            # no limitations on min/max date essentially
-            contact_type = contact_type,
-            demographic = c("Age10", "Sex", "Indigenous"),
-            measure = "Influenza", type_diabetes = TRUE,
-            ignoreOld = TRUE, lazy = FALSE, store = FALSE
-          ) %>>%
-            dplyr::mutate(QIM = "QIM 05",
-                          Measure = "InfluenzaDone",
-                          DateFrom = as.character(date_from),
-                          DateTo = as.character(date_to)) %>>%
-            dplyr::rename(State = InfluenzaDone,
-                          ProportionDemographic = Proportion_Demographic) %>>%
-            dplyr::select(QIM, Age10, Sex, Indigenous, DiabetesType,
-                          Measure, State, n, ProportionDemographic,
-                          DateFrom, DateTo)
-          # keep Age10, Sex, Indigenous, DiabetesType, HbA1CDone,
-          # n, Proportion_Demographic
-          report_values(
-            rbind(
-              report_values(),
-              qim05
-            )
+          qim05 <- getReport(
+            dMQIM$report_qim_diabetes,
+            progress_detail = "QIM 05 - Diabetes Influenza",
+            measure = "Influenza", require_type_diabetes = TRUE,
+            qim_name = "QIM 05", measure_name = "InfluenzaDone",
+            state_variable_name = dplyr::quo(InfluenzaDone)
           )
+          qim <- rbind(qim, qim05)
         }
 
         if (measure_names[[6]] %in% input$report_qim_chosen) {
-          progress$inc(amount = 1, detail = "QIM 06 - COPD influenza")
-          qim06 <- dMQIM$report_qim_copd(
-            contact = TRUE, date_from = date_from, date_to = date_to,
-            min_contact = min_contact,
-            # 'default' is clinician list chosen in right panel
-            min_date = as.Date("2000-01-01"), max_date = Sys.Date(),
-            # no limitations on min/max date essentially
-            contact_type = contact_type,
-            demographic = c("Age10", "Sex", "Indigenous"),
-            ignoreOld = TRUE, lazy = FALSE, store = FALSE
-          ) %>>%
-            dplyr::mutate(QIM = "QIM 06",
-                          Measure = "InfluenzaDone",
-                          DiabetesType = "",
-                          DateFrom = as.character(date_from),
-                          DateTo = as.character(date_to)) %>>%
-            dplyr::rename(State = InfluenzaDone,
-                          ProportionDemographic = Proportion_Demographic) %>>%
-            dplyr::select(QIM, Age10, Sex, Indigenous, DiabetesType,
-                          Measure, State, n, ProportionDemographic,
-                          DateFrom, DateTo)
-          # keep Age10, Sex, Indigenous, DiabetesType, HbA1CDone,
-          # n, Proportion_Demographic
-          report_values(
-            rbind(
-              report_values(),
-              qim06
-            )
+          qim06 <- getReport(
+            dMQIM$report_qim_copd,
+            progress_detail = "QIM 06 - COPD Influenza",
+            measure = NA, require_type_diabetes = FALSE,
+            qim_name = "QIM 06", measure_name = "InfluenzaDone",
+            state_variable_name = dplyr::quo(InfluenzaDone)
           )
+          qim <- rbind(qim, qim06)
         }
 
         if (measure_names[[7]] %in% input$report_qim_chosen) {
-          progress$inc(amount = 1, detail = "QIM 07 - 15+ Alcohol")
-          qim07 <- dMQIM$report_qim_15plus(
-            contact = TRUE, date_from = date_from, date_to = date_to,
-            min_contact = min_contact,
-            # 'default' is clinician list chosen in right panel
-            min_date = as.Date("2000-01-01"), max_date = Sys.Date(),
-            # no limitations on min/max date essentially
-            contact_type = contact_type,
-            demographic = c("Age10", "Sex", "Indigenous"),
-            measure = "Alcohol",
-            ignoreOld = TRUE, lazy = FALSE, store = FALSE
-          ) %>>%
-            dplyr::mutate(QIM = "QIM 07",
-                          Measure = "AlcoholDone",
-                          DiabetesType = "",
-                          DateFrom = as.character(date_from),
-                          DateTo = as.character(date_to)) %>>%
-            dplyr::rename(State = AlcoholDone,
-                          ProportionDemographic = Proportion_Demographic) %>>%
-            dplyr::select(QIM, Age10, Sex, Indigenous, DiabetesType,
-                          Measure, State, n, ProportionDemographic,
-                          DateFrom, DateTo)
-          # keep Age10, Sex, Indigenous, DiabetesType, HbA1CDone,
-          # n, Proportion_Demographic
-          report_values(
-            rbind(
-              report_values(),
-              qim07
-            )
+          qim07 <- getReport(
+            dMQIM$report_qim_15plus,
+            progress_detail = "QIM 07 - 15+ Alcohol",
+            measure = "Alcohol", require_type_diabetes = FALSE,
+            qim_name = "QIM 07", measure_name = "AlcoholDone",
+            state_variable_name = dplyr::quo(AlcoholDone)
           )
+          qim <- rbind(qim, qim07)
         }
 
         if (measure_names[[8]] %in% input$report_qim_chosen) {
-          progress$inc(amount = 1, detail = "QIM 08 - CVD Risk")
-          qim08 <- dMQIM$report_qim_cvdRisk(
-            contact = TRUE, date_from = date_from, date_to = date_to,
-            min_contact = min_contact,
-            # 'default' is clinician list chosen in right panel
-            min_date = as.Date("2000-01-01"), max_date = Sys.Date(),
-            # no limitations on min/max date essentially
-            contact_type = contact_type,
-            demographic = c("Age10", "Sex", "Indigenous"),
-            ignoreOld = TRUE, lazy = FALSE, store = FALSE
-          ) %>>%
-            dplyr::mutate(QIM = "QIM 08",
-                          Measure = "CVD Risk Done",
-                          DiabetesType = "",
-                          DateFrom = as.character(date_from),
-                          DateTo = as.character(date_to)) %>>%
-            dplyr::rename(State = CVDriskDone,
-                          ProportionDemographic = Proportion_Demographic) %>>%
-            dplyr::select(QIM, Age10, Sex, Indigenous, DiabetesType,
-                          Measure, State, n, ProportionDemographic,
-                          DateFrom, DateTo)
-          # keep Age10, Sex, Indigenous, DiabetesType, HbA1CDone,
-          # n, Proportion_Demographic
-          report_values(
-            rbind(
-              report_values(),
-              qim08
-            )
+          qim08 <- getReport(
+            dMQIM$report_qim_cvdRisk,
+            progress_detail = "QIM 08 - CVD Risk",
+            measure = NA, require_type_diabetes = FALSE,
+            qim_name = "QIM 08", measure_name = "CVD Risk Done",
+            state_variable_name = dplyr::quo(CVDriskDone)
           )
+          qim <- rbind(qim, qim08)
         }
 
         if (measure_names[[9]] %in% input$report_qim_chosen) {
-          progress$inc(amount = 1, detail = "QIM 09 - Cervical Screening")
-          qim09 <- dMQIM$report_qim_cst(
-            contact = TRUE, date_from = date_from, date_to = date_to,
-            min_contact = min_contact,
-            # 'default' is clinician list chosen in right panel
-            min_date = as.Date("2000-01-01"), max_date = Sys.Date(),
-            # no limitations on min/max date essentially
-            contact_type = contact_type,
-            demographic = c("Age10", "Sex", "Indigenous"),
-            ignoreOld = TRUE, lazy = FALSE, store = FALSE
-          ) %>>%
-            dplyr::mutate(QIM = "QIM 09",
-                          Measure = "CST Done",
-                          DiabetesType = "",
-                          DateFrom = as.character(date_from),
-                          DateTo = as.character(date_to)) %>>%
-            dplyr::rename(State = CSTDone,
-                          ProportionDemographic = Proportion_Demographic) %>>%
-            dplyr::select(QIM, Age10, Sex, Indigenous, DiabetesType,
-                          Measure, State, n, ProportionDemographic,
-                          DateFrom, DateTo)
-          # keep Age10, Sex, Indigenous, DiabetesType, HbA1CDone,
-          # n, Proportion_Demographic
-          report_values(
-            rbind(
-              report_values(),
-              qim09
-            )
+          qim09 <- getReport(
+            dMQIM$report_qim_cst,
+            progress_detail = "QIM 09 - Cervical screening",
+            measure = NA, require_type_diabetes = FALSE,
+            qim_name = "QIM 09", measure_name = "CST Done",
+            state_variable_name = dplyr::quo(CSTDone)
           )
-        }
+          qim <- rbind(qim, qim09)        }
 
         if (measure_names[[10]] %in% input$report_qim_chosen) {
-          progress$inc(amount = 1, detail = "QIM 10 - Diabetes BP")
-          qim10 <- dMQIM$report_qim_diabetes(
-            contact = TRUE, date_from = date_from, date_to = date_to,
-            min_contact = min_contact,
-            # 'default' is clinician list chosen in right panel
-            min_date = as.Date("2000-01-01"), max_date = Sys.Date(),
-            # no limitations on min/max date essentially
-            contact_type = contact_type,
-            demographic = c("Age10", "Sex", "Indigenous"),
-            measure = "BP", type_diabetes = TRUE,
-            ignoreOld = TRUE, lazy = FALSE, store = FALSE
-          ) %>>%
-            dplyr::mutate(QIM = "QIM 10",
-                          Measure = "BPDone",
-                          DateFrom = as.character(date_from),
-                          DateTo = as.character(date_to)) %>>%
-            dplyr::rename(State = BPDone,
-                          ProportionDemographic = Proportion_Demographic) %>>%
-            dplyr::select(QIM, Age10, Sex, Indigenous, DiabetesType,
-                          Measure, State, n, ProportionDemographic,
-                          DateFrom, DateTo)
-          # keep Age10, Sex, Indigenous, DiabetesType, HbA1CDone,
-          # n, Proportion_Demographic
-          report_values(
-            rbind(
-              report_values(),
-              qim10
-            )
+          qim10 <- getReport(
+            dMQIM$report_qim_diabetes,
+            progress_detail = "QIM 10 - Diabetes BP",
+            measure = "BP", require_type_diabetes = TRUE,
+            qim_name = "QIM 10", measure_name = "BP Done",
+            state_variable_name = dplyr::quo(BPDone)
           )
+          qim <- rbind(qim, qim10)
         }
 
+        report_values(
+          rbind(
+            report_values(),
+            qim %>>%
+              dplyr::arrange(QIM, Age10, Sex, Indigenous, DiabetesType, State) %>>%
+              dplyr::mutate(
+                DateFrom = date_from, DateTo = date_to,
+                ContactType = paste(contact_type, collapse = ", "),
+                MinContact = min_contact,
+                Clinicians = paste(dMQIM$dM$clinicians, collapse = ", ")
+              )
+          )
+        )
+        # close QIM progress bar
+        progress$close()
+
         if (input$report_number > 1) {
-        # update progress bar
+          # update progress bar
           report_progress$inc(
             amount = 1,
             detail = as.character(min(i + 1, input$report_number))
           )
-        }
-        # close QIM progress bar
-        progress$close()
-
-        report_spacing <- paste0(
-          "-", as.numeric(input$report_spacing_n),
-          " ",
-          stringi::stri_sub(
-            tolower(input$report_spacing_unit), 1, -2
+          # update dates for next loop
+          report_spacing <- paste0(
+            "-", as.numeric(input$report_spacing_n),
+            " ",
+            stringi::stri_sub(
+              tolower(input$report_spacing_unit), 1, -2
+            )
           )
-        )
-        date_from <- seq.Date(
-          date_from,
-          by = report_spacing,
-          length.out = 2
-        )[[2]]
-        date_to <- seq.Date(
-          date_to,
-          by = report_spacing,
-          length.out = 2
-        )[[2]]
+          date_from <- seq.Date(
+            date_from,
+            by = report_spacing,
+            length.out = 2
+          )[[2]]
+          date_to <- seq.Date(
+            date_to,
+            by = report_spacing,
+            length.out = 2
+          )[[2]]
+        }
       }
     })
 
@@ -651,6 +537,8 @@ qim_reportCreator <- function(input, output, session, dMQIM) {
 
   # warn if non-standard 'contact' type and number being chosen
   contact_type_warning <- shiny::reactiveVal(FALSE)
+  contact_appointment_warning <- shiny::reactiveVal(FALSE)
+  contact_visits_warning <- shiny::reactiveVal(FALSE)
   shiny::observeEvent(
     c(input$report_contact_type, input$report_min_contact),
     ignoreInit = TRUE, ignoreNULL = FALSE, {
@@ -668,8 +556,37 @@ qim_reportCreator <- function(input, output, session, dMQIM) {
         )
         contact_type_warning(TRUE)
       }
+
+      if ("Appointments" %in% input$report_contact_type &&
+          !contact_appointment_warning()) {
+        shinytoastr::toastr_warning(
+          message = paste(
+            "Choose valid appointment 'status' from the right sidebar.",
+            "e.g. Invoiced, Waiting, Booked..."
+          ),
+          position = "bottom-center",
+          closeButton = TRUE,
+          timeOut = 20000
+        )
+        contact_appointment_warning(TRUE)
+      }
+
+      if ("Visits" %in% input$report_contact_type &&
+          !contact_visits_warning()) {
+        shinytoastr::toastr_warning(
+          message = paste(
+            "Choose valid visit 'types' from the right sidebar.",
+            "e.g. Surgery, Telephone, Telhealth, SMS, Email..."
+          ),
+          position = "bottom-center",
+          closeButton = TRUE,
+          timeOut = 20000
+        )
+        contact_visits_warning(TRUE)
+      }
     }
   )
+
 
   # warn if more than one report being generated
   # could take a long time!